updata portM0, M3

+++ sources/kernel/mthread/inc/mutex.h
@@ -8,7 +8,7 @@ typedef void *descriptorTask;
 
 struct Mmutex {
     /// Блокировка мьютекса
-    uint8_t lock;
+    uint32_t lock;
     /// Дескриптор потока занявшего мьютекс
     descriptorTask pTask_Descriptor;
     /// Запрос на блокировку мьютекса
